fix crc check, fix hmac key creation
